week_4/src/requirements.txt
torch
torchvision
tqdm
Pillow
numpy